+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [4.2.1] - 22 November 2023
+- Fix `SkadiSimpleDialog` content cutout on iOS
+- Add `error` constructor to `SkadiSimpleDialog`
+- New extension
+  -  `rfs` : **TextStyle**
+  -  
 ## [4.2.0] - 7 November 2023
 - add `titleBuilder` for `SkadiAccordion`